玉が転がっていくサンプルを実装
